workbench changes for: Moving API YAML files and common ruby code to apidef repo.
