+++ api/build.gradle
@@ -2,7 +2,7 @@ import io.swagger.codegen.config.CodegenConfigurator
 import io.swagger.codegen.DefaultGenerator
 
 def swaggerTemplateDir = 'src/main/resources'
-def swaggerSourceFile = 'src/main/resources/workbench.yaml'
+def swaggerSourceFile = '../apidef/workbench.yaml'
 def swaggerTargetFolder = 'src/generated/java'
 
 task generateApi {
