+++ public-api/build.gradle
@@ -2,7 +2,7 @@ import io.swagger.codegen.config.CodegenConfigurator
 import io.swagger.codegen.DefaultGenerator
 
 def swaggerTemplateDir = 'src/main/resources'
-def swaggerSourceFile = 'src/main/resources/public-api.yaml'
+def swaggerSourceFile = '../apidef/public-api.yaml'
 def swaggerTargetFolder = 'src/generated/java'
 
 task generateApi {
