+++ ui/package.json
@@ -11,9 +11,9 @@
     "e2e": "ng e2e",
     "codegen": "yarn run codegen-clean && yarn run codegen-swagger && yarn run post-codegen && yarn run public-codegen-swagger && yarn run public-post-codegen",
     "codegen-clean": "rm -rf src/generated src/publicGenerated",
-    "codegen-swagger": "java -jar ../libproject/swagger-codegen-cli.jar generate --lang typescript-angular --input-spec ../api/src/main/resources/workbench.yaml --output src/generated --additional-properties ngVersion=2",
+    "codegen-swagger": "java -jar ../libproject/swagger-codegen-cli.jar generate --lang typescript-angular --input-spec ../apidef/workbench.yaml --output src/generated --additional-properties ngVersion=2",
     "post-codegen": "sed -i.bak -e 's/OpaqueToken/InjectionToken/' src/generated/variables.ts",
-    "public-codegen-swagger": "java -jar ../libproject/swagger-codegen-cli.jar generate --lang typescript-angular --input-spec ../public-api/src/main/resources/public-api.yaml --output src/publicGenerated --additional-properties ngVersion=2",
+    "public-codegen-swagger": "java -jar ../libproject/swagger-codegen-cli.jar generate --lang typescript-angular --input-spec ../apidef/public-api.yaml --output src/publicGenerated --additional-properties ngVersion=2",
     "public-post-codegen": "sed -i.bak -e 's/OpaqueToken/InjectionToken/' src/publicGenerated/variables.ts"
 
   },
