+++ ui/project.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-require_relative "../libproject/workbench"
+require_relative "../apidef/libproject/workbench"
 require_relative "libproject/devstart.rb"
 
 Workbench.handle_argv_or_die(__FILE__)
